+++ packages/website-generator/src/index.test.ts
@@ -1,6 +1,0 @@
-import test from 'tape';
-
-test('website-generator', (t: test.Test) => {
-  // @TODO: write tests for the entry function.
-  t.end();
-});